training-solutions/src/main/java/methodparams/marriage/Marriage.java
package methodparams.marriage;

import java.time.LocalDate;

public class Marriage {
    public void getMarried(Woman woman, Man man){
        setNameAfterMarriage(woman,man);
        addRegisterDate(woman,man);
    }

    private void setNameAfterMarriage(Woman woman,Man man){
        woman.setName(man.getName().split(" ")[0]+"né" + woman.getName());
    }

    private void addRegisterDate(Woman woman, Man man) {
        woman.addRegisterDate(new RegisterDate("házasságkötés ideje", LocalDate.now()));
        man.addRegisterDate(new RegisterDate("házasságkötés ideje", LocalDate.now()));
    }


}
